ISSUE#28 Обьединение контроллеров
